+++ history.md
@@ -1,11 +1,12 @@
 
-### 2.0.19ing
+### 2.0.19
 - 测试结构更新
 - formatTree逻辑优化，删除formatTreeNext,测试用例添加
 - updateList中option.format不返回需要push的数据，因为此时需要的对象都是复杂格式，内存指针形式，因此format仅对对象数据做格式化，返回值为是否添加到数组中
 - updateDataWidthOption实现total模式
 - updateDataWidthOption参数顺序调整
 - 添加isSymbol/formatNum，优化getType的symbol判断
+- getType逻辑优化，二参数为真时进行额外判断，基础判断添加array，基础中不对symbol做额外的可能性判断
 
 ### 2.0.18
 - current注释